mad science DSL research.
Tweaking the AST in various interesting ways for fun and profit:
it's possible, and it may even be neat.

Snapshot commit.  All tests are completely broken; and
debug printfs are strewn everywhere.  I'm still making a checkpoint
because I'm afraid to break something along the way when cleaning it
up into real stable working things (and, I suspect I might want the
comments for later review, to see if I missed any angles).