Refactor internals.  Attempt to document them better.
With partial success.  Some of my attempts to separate the phases of
exploration and evaluation ran aground on the fact I'm just mutating
the AST, and if there's a deep-copy feature for the AST, then I haven't
stumbled across it yet.  (I really hope it's going to be there if I
just look again; it would be so very useful.  Surely I'm just not
looking in the right place.)

Still.  Docs got better.

Printouts on the debug channel got some more decoration saying
where they're coming from.  This affects the current test fixtures,
so they're regenerated as well.

Error handling started to get a _bit_ structured.
I started adding a "wfxapi" package for API things.
Which is, so far, entirely errors.  Heh.
Much more to come on that still, I'm sure.

Incidentally, I've come to notice the tests pass sporadically.
The toposort library I reached for doesn't support stability yet :(
But it's being discussed upstream and they do want to add that feature,
so I'll attempt to ignore the trouble for now.