Don't attach causes in exec errors; unhelpful.
The values we were attaching as cause were the golang stdlib errors,
and A) stringifying those doesn't help much; we already transcribed
pretty much everything it would say, and B) go-serum "helpfully" adds
more description of the golang type, in its eagerness to be producing
info that's host-language-agnostic... but really, that's just noise
in this situation, considering (A).  (And realistically, probably a
nudge to make you not cause-chain to non-serum error values anyway.)
So: all in all, chaining those values around was unhelpful.

Also use the |q formatting in signal errors too.