Work towards standardizing the error codes a bit more.
Gonna continue incrementally nudging this over time.
I don't yet know what the most usable groupings are going to be.

At the very least, now it's decided that they're going in the 'wfxapi'
package, which lets me start start using constants, and start removing
the TODOs about package cycles!

There's also going to be a big harsh utility limit here because
a huge amount of control flow is bouncing through the starlark
interpreter, and that is most DEFINITELY the kind of thing that
static analysis tools aren't going to be able to help us trace much!
If, indeed, at all.

But we can still get things right up to that edge being analyzed.

And then we'll just have to filter the things it produces again.
No real way around it.

Perfect can't be allowed to be the enemy of good, I guess.